--- src/main/java/goormthonuniv/team_22_be/questioncard/infrastructure/AnswerRepository.java
@@ -1,8 +1,0 @@
-package goormthonuniv.team_22_be.questioncard.infrastructure;
-
-import goormthonuniv.team_22_be.questioncard.domain.model.Answer;
-import goormthonuniv.team_22_be.questioncard.domain.repository.AnswerRepositoryCustom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AnswerRepository extends JpaRepository<Answer, Long>, AnswerRepositoryCustom {
-}
